+++ src/main/java/cz/aegis/corebanking/controller/AccountController.java
@@ -2,13 +2,10 @@ package cz.aegis.corebanking.controller;
 
 import cz.aegis.corebanking.dto.AccountResponse;
 import cz.aegis.corebanking.dto.CreateAccountRequest;
-import cz.aegis.corebanking.dto.DepositMoneyRequest;
-import cz.aegis.corebanking.dto.WithdrawMoneyRequest;
 import cz.aegis.corebanking.service.AccountService;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -31,23 +28,5 @@ public class AccountController {
         AccountResponse response = accountService.createAccount(request);
 
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
-    }
-
-    //endpoint POST - deposit
-    @PostMapping("/{id}/deposit")
-    public ResponseEntity<AccountResponse> depositMoney(@PathVariable Long id, @RequestBody @Valid DepositMoneyRequest request) {
-
-        AccountResponse response = accountService.depositMoney(id, request);
-
-        return ResponseEntity.ok(response);
-    }
-
-    //endpoint POST - withdraw
-    @PostMapping("/{id}/withdraw")
-    public ResponseEntity<AccountResponse> withdrawMoney(@PathVariable Long id, @RequestBody @Valid WithdrawMoneyRequest request) {
-
-        AccountResponse response = accountService.withdrawMoney(id, request);
-
-        return ResponseEntity.ok(response);
     }
 }