+++ src/main/java/cz/aegis/corebanking/exception/InvalidTransactionTypeException.java
@@ -1,0 +1,8 @@
+package cz.aegis.corebanking.exception;
+
+public class InvalidTransactionTypeException extends RuntimeException{
+
+    public InvalidTransactionTypeException(String transactionType) {
+        super("Invalid transaction type: " + transactionType);
+    }
+}